requirements.txt
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
pyyaml==6.0.2

mlflow==2.16.2

fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2

python-dotenv==1.0.1
joblib==1.4.2

matplotlib==3.9.2